go schemabuilder/output: Move output/result specific code to separate file
Summary: consolidates the code for managing the output/response type
conversions for a graphql query into it's own file.
